% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_compare.R
\name{svy_compare}
\alias{svy_compare}
\title{svy_compare}
\usage{
svy_compare(variables, data)
}
\arguments{
\item{variables}{Vector of columns from data.}

\item{data}{A data object that can be either a survey object or other data type.}
}
\value{
A data.frame with counts of each possible combination.
}
\description{
This is a helper function to get counts of how many times each combination of levels of a provided vector of variables appears in the provided dataset. Useful for tracking programmatic logic.
}
