% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_prop.R
\name{svy_prop}
\alias{svy_prop}
\title{svy_prop}
\usage{
svy_prop(row, column, data)
}
\arguments{
\item{row}{Row variable from survey object.}

\item{column}{Column variable from survey object.}

\item{data}{Survey object.}
}
\value{
A dataframe with survey frequencies put into a table of proportions.
}
\description{
This is helper function to create a table of survey frequency proportions with one row variable and one column variable.
}
\examples{
svy_prop(row = "b1a", column = "raceth", data = svy_df)
}
